# ts-backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "fulcrumaxe ts-backend API",
    "version": "0.1.0",
    "description": "TypeScript/Bun+Hono backend — mirrors the Python backend with 1:1 parity. Runs on 127.0.0.1:19099 (loopback-only). Auth: set AF_API_AUTH_KEY to enable; unset = auth disabled. SSE routes (/feed, /events) accept ?token= for EventSource clients that cannot send custom headers. See https://github.com/fulcrumaxe/fulcrumaxe/discussions/1437 for the full phased migration plan.",
    "contact": {
      "url": "https://github.com/fulcrumaxe/fulcrumaxe"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:19099",
      "description": "Default loopback server (TS_BACKEND_PORT overrides 19099)"
    }
  ],
  "security": [],
...
  "paths": {
...
    "/budget/init": {
      "post": {
        "summary": "Initialize budget session",
        "description": "Writes three file-based blackboard entries: budget/session_ceiling, budget/session_spent (0), and budget/per_agent_ceiling. Python remains writer of record on the production blackboard; this route is loopback-only and parity-gated.",
        "operationId": "initBudget",
        "tags": [
          "Budget"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BudgetInitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Budget initialized successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetInitResponse"
                }
              }
            }
          },
          "400": {
            "description": "'ceiling' present but not a positive integer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          },
          "401": {
            "description": "No credentials provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid token or RBAC denial",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          }
        }
      }
    },
...
    "/graphql": {
      "post": {
        "summary": "Home-grown GraphQL",
        "description": "Hand-rolled GraphQL endpoint with 1:1 parity to the Python graphql_api.py implementation. Supported root query fields: health, budget, cost, registry, agents, kpi, control, audit, replays, spawnQueue, notifications, plugins. __schema and __type introspection are supported. Auth: bearer token + RBAC. Note: several fields are always null due to documented key-name mismatches in the Python reference (budget.used, cost.total_usd, registry.stats.open, etc.).",
        "operationId": "graphql",
        "tags": [
          "GraphQL"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GraphqlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "GraphQL response. data may coexist with errors when partial results are available.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphqlResponse"
                }
              }
            }
          },
          "400": {
            "description": "'query' field missing from request body",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          },
          "401": {
            "description": "No credentials provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          },
          "403": {
            "description": "Invalid token or RBAC denial",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetailError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token from AF_API_AUTH_KEY. Required when auth is enabled. SSE routes (/feed, /events) also accept ?token=<key> as a query param because EventSource clients cannot set custom headers."
      }
    },
    "schemas": {
      "DetailError": {
        "type": "object",
        "required": [
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable error message"
          }
        }
      },
...
      "BudgetStatus": {
        "type": "object",
        "required": [
          "ceiling",
          "spent",
          "remaining",
          "per_agent_ceiling",
          "warn_threshold_pct",
          "agents"
        ],
        "properties": {
          "ceiling": {
            "type": "number",
            "example": 5000000
          },
          "spent": {
            "type": "number",
            "example": 0
          },
          "remaining": {
            "type": "number",
            "example": 5000000
          },
          "per_agent_ceiling": {
            "type": "number",
            "example": 500000
          },
          "warn_threshold_pct": {
            "type": "number",
            "example": 80
          },
          "agents": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": true
            },
            "description": "Per-agent budget usage records"
          }
        }
      },
      "BudgetInitRequest": {
        "type": "object",
        "properties": {
          "ceiling": {
            "type": "integer",
            "description": "Session token ceiling. Must be a positive integer. Omit to use the configured default (5,000,000).",
            "example": 5000000
          }
        }
      },
      "BudgetInitResponse": {
        "type": "object",
        "required": [
          "ok",
          "status"
        ],
        "properties": {
          "ok": {
            "type": "boolean",
            "example": true
          },
          "status": {
            "$ref": "#/components/schemas/BudgetStatus"
          }
        }
      },
...
      "GraphqlRequest": {
        "type": "object",
        "required": [
          "query"
        ],
        "properties": {
          "query": {
            "type": "string",
            "description": "GraphQL query string. Supported root fields: health, budget, cost, registry, agents, kpi, control, audit, replays, spawnQueue, notifications, plugins. __schema and __type introspection are supported.",
            "example": "{ health { ok loop { healthy } } }"
          }
        }
      },
      "GraphqlResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": true,
            "description": "Present when the query executed (may coexist with errors)"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "message"
              ],
              "properties": {
                "message": {
                  "type": "string"
                },
                "path": {
                  "type": "string"
                }
              }
            },
            "description": "Present when one or more field errors occurred"
          }
        }
      },
...
    }
  }
}